requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
arrow==1.4.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
deprecation==2.1.0
distro==1.9.0
fastapi==0.124.4
fastkml==1.4.0
gotrue==2.12.4
groq==0.32.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
lxml==6.0.2
multidict==6.7.0
numpy==2.4.0
packaging==25.0
postgrest==2.25.1
propcache==0.4.1
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pygeoif==1.6.0
PyJWT==2.10.1
pyshp==3.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
realtime==2.25.1
requests==2.32.5
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
storage3==2.25.1
StrEnum==0.4.15
supabase==2.25.1
supabase-auth==2.25.1
supabase-functions==2.25.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.38.0
websockets==15.0.1
yarl==1.22.0